react-hitl-mcp-agent/requirements.txt
# Core agent stack
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0   # AsyncSqliteSaver — bundled with LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# LLM providers (install whichever you use)
langchain-groq>=0.2.0      # Groq (llama-3.3-70b-versatile)
# langchain-anthropic      # Claude — uncomment if needed

# MCP integration
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0

# Structured LLM reasoning (guardrails + eligibility assessment)
dspy-ai>=2.5.0

# API server (optional — needed only for the HTTP API, not the CLI demo)
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
